Improve treasure selection ui
